requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

# Apify SDK
apify==1.1.3
apify-client==1.4.0

# LangGraph and related packages
langgraph==0.0.20

# OpenAI
openai>=1.10.0

# Utilities
python-dotenv==1.0.0
pydantic==1.10.8
tenacity==8.2.3

# Web scraping
beautifulsoup4==4.12.2
httpx==0.24.1